+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to `laravel-livewire-tables` will be documented in this file
 
 ## [Unreleased]
 
+## [2.7.0] - 2022-05-07
+
+### Added
+
+- Added functionality to hide individual filters from popover and slide down views
+- Added functionality to hide individual filters from filter pills
+- Added functionality to hide individual filters from the active filter count
+- Added functionality to say which filters get reset by the clear button
+- Added functionality to set filters as secondaryHeader or footer of columns
+
 ## [2.6.0] - 2022-05-05
 
 ### Added
@@ -659,7 +669,8 @@ Ground Up Rebuild
 
 - Initial release
 
-[Unreleased]: https://github.com/rappasoft/laravel-livewire-tables/compare/v2.6.0...development
+[Unreleased]: https://github.com/rappasoft/laravel-livewire-tables/compare/v2.7.0...development
+[2.7.0]: https://github.com/rappasoft/laravel-livewire-tables/compare/v2.6.0...v2.7.0
 [2.6.0]: https://github.com/rappasoft/laravel-livewire-tables/compare/v2.5.0...v2.6.0
 [2.5.0]: https://github.com/rappasoft/laravel-livewire-tables/compare/v2.4.0...v2.5.0
 [2.4.0]: https://github.com/rappasoft/laravel-livewire-tables/compare/v2.3.0...v2.4.0